added atomic design to Map page
