+++ package.json
@@ -20,7 +20,8 @@
     "react-native-screens": "^2.0.0-beta.4",
     "react-navigation": "^4.1.1",
     "react-navigation-stack": "^2.1.1",
-    "react-navigation-tabs": "^2.7.0"
+    "react-navigation-tabs": "^2.7.0",
+    "styled-components": "^5.0.1"
   },
   "devDependencies": {
     "@babel/core": "^7.8.4",
